MUST be zipped and SHOULD have .crate.zip?
I don't think we should be picky about the filename,
as anyone consuming a zip can simply look for ro-crate-metadata.json inside
and thus determine it is an RO-Crate.

I think it is true you want the ro-crate-metadata.json straight in the root, and
do not like somefolder/ro-crate-metadata.json?

.jsonld -> .json